==> etl_proces/workflow-notebooks/2_poloha.ipynb <==
import geopandas as gpd

obec_okres_kraj = gpd.read_file(r'./source-files/obec_okres_kraj.geojson')
sportoviska = gpd.read_file(r'./source-files/1_klasifikovane.geojson')

print(obec_okres_kraj.crs)
print(sportoviska.crs)

# Priradi Obec (s tym okres aj kraj) tak, ze pre kazde sportovisko spocita obsah pretinania geometrie sportoviska a geometrie obce a priradi sportovisko tej obci, s ktorou ma najvacsi prienik

poloha_data = gpd.sjoin(sportoviska, obec_okres_kraj, how="inner", predicate="intersects")

poloha_data['intersection_area'] = poloha_data.apply(
    lambda row: row['geometry'].intersection(obec_okres_kraj.loc[row['index_right'], 'geometry']).area,
    axis=1
)

poloha_data.reset_index(inplace=True)
poloha_data = poloha_data.loc[poloha_data.groupby('id')['intersection_area'].idxmax()]
poloha_data = poloha_data.drop(['index_right', 'obec', 'okres', 'kraj', 'intersection_area'], axis = 1)
poloha_data

# Ak dojde k duplicite zaznamu o sportovisku -> nieco je zle -> vyhodi error

counts = poloha_data.groupby('id').size()
counts = counts[counts >= 2]

if not counts.empty:
    raise Exception('Pri priradovani obce doslo k zdvojeniu zaznamu o sportovisku')

# kontroluje ci vsetkym sportoviskam bola priradena obec, ak nie vyhodi chybu
filtered_df = poloha_data.dropna(subset=['obec_id', 'okres_id', 'kraj_id'])
removed_rows = poloha_data[~poloha_data.index.isin(filtered_df.index)]
if removed_rows.shape[0] > 0:
    raise Exception('Nejakym sportoviskam sa nepodarilo priradit obec')

poloha_data.to_file(r'./source-files/2_doplnena_obec.geojson', driver = 'GeoJSON')
