==> etl_proces/workflow-notebooks/4_nacitaj_do_db.ipynb <==
import geopandas as gpd
from sqlalchemy import create_engine, text
from geoalchemy2 import Geometry, WKTElement
from sqlalchemy.orm import sessionmaker
from sqlalchemy.exc import SQLAlchemyError
import db_credentials

NAČÍTANIE DÁT

data_input = gpd.read_file(r'./source-files/3_prelozene.geojson')

data_input

PRIPOJENIE K DATABÁZE

host = db_credentials.host
dbname = db_credentials.dbname
user = db_credentials.user
password = db_credentials.password
port = 5432
table_name = 'sportoviska_source'

engine = create_engine(f'postgresql://{user}:{password}@{host}:{port}/{dbname}')

NAČÍTANIE ŠPORTOVÍSK, OBCÍ, OKRESOV A KRAJOV DO DB

data = data_input.copy()

data['geom'] = data['geometry'].apply(lambda x: WKTElement(x.wkt, srid=4326))
data.drop('geometry', axis=1, inplace=True)

# Vytvori session
Session = sessionmaker(bind=engine)
session = Session()

# V ramci session sa pokusi nahrat DataFrame do databazy
try:
    data.to_sql(table_name, engine, if_exists='replace', index=False,
                dtype={'geom': Geometry(geometry_type='GEOMETRY', srid=4326)})

    # v pripade uspenseho nahrania sa commitnu transakcie prevedene pocas sessionu
    session.commit()

except SQLAlchemyError as e:
    # v pripade neuspesneho nahrania sa transakcie prevedene pocas sessionu vratia naspat a nasledne sa session ukonci
    session.rollback()
    session.close()
    raise (f"Error pri nacitani sportovisk do databazy: {e}")

obec_okres_kraj = gpd.read_file(r'./source-files/obec_okres_kraj.geojson', driver = 'GeoJSON')
obec_okres_kraj['geom'] = obec_okres_kraj['geometry'].apply(lambda x: WKTElement(x.wkt, srid=4326))
obec_okres_kraj.drop('geometry', axis=1, inplace=True)

Session = sessionmaker(bind=engine)
session = Session()

try:
       obec_okres_kraj.to_sql('obec_okres_kraj', engine, if_exists='replace', index=False,
                       dtype={'geom': Geometry(geometry_type='GEOMETRY', srid=4326),
                              })
       session.commit()

except SQLAlchemyError as e:
       session.rollback()
       session.close()
       raise (f"Error pri nacitani obci do databazy: {e}")

EXEKÚCIA POTREBNÝCH SQL SKRIPTOV

def execute_sql_script(conn, script_path):
    """
    Načíta SQL skript zo súboru a spustí ho na databáze.

    Argumenty:
    conn (sqlalchemy.engine.Connection): Pripojenie k databáze, na ktorej sa má spustiť SQL skript.
    script_path (str): Cesta k súboru obsahujúcemu SQL skript.

    Návratová hodnota:
    None

    Príklad použitia:
    >>> engine = create_engine('postgresql://user:password@localhost/mydatabase')
    >>> conn = engine.connect()
    >>> execute_sql_script(conn, 'path/to/script.sql')
    """
    # Načíta obsah SQL skriptu zo súboru
    with open(script_path, 'r', encoding="utf8") as file:
        sql_script = file.read()

    # Vykoná SQL skript na pripojenej databáze
    conn.execute(text(sql_script))


Session = sessionmaker(bind=engine)
session = Session()

try:
    with engine.connect() as conn:
        with conn.begin():
            rename_column = text("""
                ALTER TABLE public.sportoviska_source
                RENAME COLUMN id TO osm_id;
            """)
            conn.execute(rename_column)

            add_column = text("""
                ALTER TABLE public.sportoviska_source
                ADD COLUMN id SERIAL;
            """)
            conn.execute(add_column)

            execute_sql_script(conn, r'./sql-scripts/creating_tables.sql')
            execute_sql_script(conn, r'./sql-scripts/fill_tables.sql')

    session.commit()

except SQLAlchemyError as e:
    session.rollback()
    session.close()
    raise (f"Error pri spustani SQL skriptov na databaze: {e}")
